Process some todos, refactor parser run type

+++ src/lexer_fsm.h
@@ -34,7 +34,7 @@ typedef struct lexer_fsm_t {
     char numeric_char_value[4]; // Stack for numeric value of char
     short numeric_char_position; // Head of stack for numeric value of char
 
-    int line; // Actual line
+    size_t line; // Actual line
 
     LexerError lexer_error; // Error code
 } LexerFSM;
